MOVED: test files into __tests__ subdir FIXED: did show 0 file length for broken symlinks IMPROVED: show symlink title when hovering a symlink FIXED: use error color for broken symlinks on selected row
